Move game state/reducer definition into slice file
