+++ src/store.ts
@@ -13,10 +13,10 @@ import {
 import { listenerMiddleware } from './app/action-listeners.js';
 import dynamicMiddlewares from './middleware/dynamic-middlewares';
 import { lazyReducerEnhancer, LazyStore } from './middleware/lazy-reducers.js';
-import type { GameState } from './reducers/game.js';
 import { configureAppStore } from './slices/app/app-module-configurator.js';
 import type { AppState, APP_SLICE_NAME } from './slices/app/app-slice.js';
 import type { AuthState } from './slices/auth/auth-slice.js';
+import type { GameState } from './slices/game/game-slice.js';
 import type { LiveState } from './slices/live/live-slice.js';
 import type { TeamState } from './slices/team/team-slice.js';
